backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
python-multipart
python-dotenv
pydantic
google-genai==1.51.0